_sysfs: fix crash when running with --help option
When running the _sysfs module with the --help option, we were getting a crash in the background thread because it was trying to write to stdout after the program had "exited".

    Fatal Python error: could not acquire lock for <_io.BufferedWriter name='<stdout>'> at interpreter shutdown, possibly due to daemon threads

    Thread 0x00007fe0aab0f700 (most recent call first):
      File "/home/david/work/ev3dev-dpkg/python-ev3dev-testfs/ev3dev/testfs/_sysfs.py", line 73 in _run
      File "/usr/lib/python3.6/threading.py", line 864 in run
      File "/usr/lib/python3.6/threading.py", line 916 in _bootstrap_inner
      File "/usr/lib/python3.6/threading.py", line 884 in _bootstrap

    Current thread 0x00007fe0acabc740 (most recent call first):
    Aborted (core dumped)

Fix it by checking that we are actually planning on mounting the file system before starting the background thread.